analytics-backend/supabase/requirements.txt
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
google-cloud-bigquery>=3.0.0
